# swagger.json
{
  "swagger": "2.0",
  "info": {
      "title": "Include Care API",
      "description": "API",
      "termsOfService": "",
      "contact":{
          "name": "include care",
          "email": "",
          "url": "http://localhost:3000"
      },
      "license": {
          "name": "",
          "url": ""
      },
      "version": "1.0.0"
  },
...
  "paths": {
...
      "/api/user": {
        "get": {
            "tags": ["User"],
            "summary": "get one user",
            "parameters": [
            ],
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                  "description": "OK",
                  "schema": {
                      "$ref": "#/definitions/User"
                  }
              },
              "400": {
                  "description": "Failed. Bad post data."
              }
          }
        },
        "put": {
          "tags": ["User"],
          "summary": "update one user",
          "parameters": [
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
                "description": "OK",
                "schema": {
                    "$ref": "#/definitions/User"
                }
            },
            "400": {
                "description": "Failed. Bad post data."
            }
          }
        },
        "delete": {
          "tags": ["User"],
          "summary": "update one user",
          "parameters": [
          ],
          "responses": {
            "200": {
                "description": "OK",
                "schema": {
                    "$ref": "#/definitions/User"
                }
            },
            "400": {
                "description": "Failed. Bad post data."
            }
          }
        }
      },
...
  },
  "definitions": {
...
    "User": {
        "type": "object",
        "properties": {
            "firstName": {
                "type": "string"
            },
            "lastName": {
                "type": "string"
            },
            "email": {
                "type": "string"
            },
            "phone": {
                "type": "string"
            },
            "password": {
              "type": "string"
            },
            "passwordReset": {
              "type": "string"
            },
            "messages": {
              "type": "array"
            },
            "jobs": {
              "type": "array"
            },
            "describeWorker": {
              "type": "string"
            },
            "county": {
              "type": "string"
            },
            "serviceWorker": {
              "type": "array"
            },
            "avatar": {
              "type": "object"
            },
            "role": {
              "type": "string"
            }
        }
    },
...
  }
}  